+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 - `linalg:` Ported `BLAS` `spr()`
 - `linalg:` Ported `Matrix` `copy()` and `toString()`, and implemented `clone()` to make matrices usable within RDD functions as a zero value parameter
+- `linalg:` Ported `RowMatrix` `computeGramianMatrix()` and `triuToFull()`
 
 ### Changed
 - `linalg:` Use native Lua 1-based `colPtrs` indexes in `SparseMatrix`